requirements.txt
jsonpickle==3.0.1
xml_marshaller==1.0.2
uvicorn[standard]==0.17.*
fastapi==0.95.1
PyYAML==6.0
msgpack==1.0.5
requests==2.29.0
fastavro==1.7.3
Pympler==1.0.1